feat: 사용자 주문 API 구현 및 주문 상세 내역 조회 E2E 테스트 추가

+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/order/OrderItemDetail.java
@@ -4,5 +4,5 @@ public record OrderItemDetail(
     Long productId,
     String productName,
     int quantity,
-    int price
+    Long price
 ) {}
